refactor - LottoNumberManager 에서 로또 번호를 생성 하던것을 LottoTicket으로 위임
